+++ test/augmentables/test_polys.py
@@ -421,6 +421,30 @@ def test_Polygon_is_out_of_image():
     assert not poly.is_out_of_image((10, 10, 3), fully=True, partly=False)
     assert poly.is_out_of_image((10, 10, 3), fully=False, partly=True)
 
+    # polygon with all corners outside of the image
+    poly = ia.Polygon([(-1.0, -1.0), (2.0, -1.0), (2.0, 2.0), (-1.0, 2.0)])
+    assert not poly.is_out_of_image((100, 100, 3), fully=True, partly=False)
+    assert poly.is_out_of_image((100, 100, 3), fully=False, partly=True)
+    assert not poly.is_out_of_image((1, 1, 3), fully=True, partly=False)
+    assert poly.is_out_of_image((1, 1, 3), fully=False, partly=True)
+    assert poly.is_out_of_image((1, 1, 3), fully=True, partly=True)
+
+    # polygon with two points
+    poly = ia.Polygon([(2.0, 2.0), (10.0, 2.0)])
+    assert not poly.is_out_of_image((100, 100, 3), fully=True, partly=False)
+    assert not poly.is_out_of_image((100, 100, 3), fully=False, partly=True)
+    assert not poly.is_out_of_image((3, 3, 3), fully=True, partly=False)
+    assert poly.is_out_of_image((3, 3, 3), fully=False, partly=True)
+    assert poly.is_out_of_image((1, 1, 3), fully=True, partly=False)
+    assert not poly.is_out_of_image((1, 1, 3), fully=False, partly=True)
+
+    # polygon with a single point
+    poly = ia.Polygon([(2.0, 2.0)])
+    assert not poly.is_out_of_image((100, 100, 3), fully=True, partly=False)
+    assert not poly.is_out_of_image((100, 100, 3), fully=False, partly=True)
+    assert poly.is_out_of_image((1, 1, 3), fully=True, partly=False)
+    assert not poly.is_out_of_image((1, 1, 3), fully=False, partly=True)
+
     poly = ia.Polygon([])
     got_exception = False
     try:
@@ -471,6 +495,52 @@ def _test_Polygon_cut_clip(func):
         [0.5, 1.0]
     ]))
     assert multipoly_clipped[0].label == "test"
+
+    # square poly with a single edge intersecting the image (issue #310)
+    poly = ia.Polygon([(-1.0, 0.0), (0.0, 0.0), (0.0, 1.0), (-1.0, 1.0)])
+    image = np.zeros((1, 1, 3), dtype=np.uint8)
+    multipoly_clipped = func(poly, image)
+    assert isinstance(multipoly_clipped, list)
+    assert len(multipoly_clipped) == 0
+
+    # square poly with a tiny area on the left image edge intersecting with
+    # the image
+    offset = 1e-4
+    poly = ia.Polygon([(-1.0, 0.0), (0.0+offset, 0.0),
+                       (0.0+offset, 1.0), (-1.0, 1.0)])
+    image = np.zeros((1, 1, 3), dtype=np.uint8)
+    multipoly_clipped = func(poly, image)
+    assert isinstance(multipoly_clipped, list)
+    assert len(multipoly_clipped) == 1
+    assert multipoly_clipped[0].exterior_almost_equals(np.float32([
+        [0.0, 0.0],
+        [0.0+offset, 0.0],
+        [0.0+offset, 1.0],
+        [0.0, 1.0]
+    ]))
+
+    # square poly with a single point intersecting the image (issue #310)
+    poly = ia.Polygon([(-1.0, -1.0), (0.0, -1.0), (0.0, 0.0), (-1.0, 0.0)])
+    image = np.zeros((1, 1, 3), dtype=np.uint8)
+    multipoly_clipped = func(poly, image)
+    assert isinstance(multipoly_clipped, list)
+    assert len(multipoly_clipped) == 0
+
+    # square poly with a tiny area around the top left image corner
+    # intersecting with the the image
+    offset = 1e-4
+    poly = ia.Polygon([(-1.0, -1.0), (0.0, -1.0),
+                       (0.0+offset, 0.0+offset), (-1.0, 0.0)])
+    image = np.zeros((1, 1, 3), dtype=np.uint8)
+    multipoly_clipped = func(poly, image)
+    assert isinstance(multipoly_clipped, list)
+    assert len(multipoly_clipped) == 1
+    assert multipoly_clipped[0].exterior_almost_equals(np.float32([
+        [0.0, 0.0],
+        [0.0+offset, 0.0],
+        [0.0+offset, 0.0+offset],
+        [0.0, 0.0+offset]
+    ]))
 
     # non-square poly, with one rectangle on the left side of the image and one on the right side,
     # both sides are connected by a thin strip below the image
@@ -529,6 +599,39 @@ def _test_Polygon_cut_clip(func):
         [100, 100],
         [0.5*100, 100]
     ]))
+
+    # polygon with two points
+    poly = ia.Polygon([(2.0, 2.0), (10.0, 2.0)])
+    multipoly_clipped = func(poly, (100, 100, 3))
+    assert isinstance(multipoly_clipped, list)
+    assert len(multipoly_clipped) == 1
+    assert multipoly_clipped[0].exterior_almost_equals(np.float32([
+        [2.0, 2.0],
+        [10.0, 2.0]
+    ]))
+
+    poly = ia.Polygon([(2.0, 2.0), (10.0, 2.0)])
+    multipoly_clipped = func(poly, (3, 3, 3))
+    assert isinstance(multipoly_clipped, list)
+    assert len(multipoly_clipped) == 1
+    assert multipoly_clipped[0].exterior_almost_equals(np.float32([
+        [2.0, 2.0],
+        [3.0, 2.0]
+    ]), max_distance=1e-3)
+
+    # polygon with a single point
+    poly = ia.Polygon([(2.0, 2.0)])
+    multipoly_clipped = func(poly, (3, 3, 3))
+    assert isinstance(multipoly_clipped, list)
+    assert len(multipoly_clipped) == 1
+    assert multipoly_clipped[0].exterior_almost_equals(np.float32([
+        [2.0, 2.0]
+    ]))
+
+    poly = ia.Polygon([(2.0, 2.0)])
+    multipoly_clipped = func(poly, (1, 1, 3))
+    assert isinstance(multipoly_clipped, list)
+    assert len(multipoly_clipped) == 0
 
 
 def test_Polygon_shift():